Add nomad optimization and fixups
