+++ .gitignore
@@ -1,2 +1,4 @@
 # OS-related files/folders
 .DS_Store
+
+data/
